MB-66105: fixup profile loading couchdb
We need access to the profiles from the context of the couchdb node (not
ns_server node) and so we need to load it so it has access to the
persistent_term database as well. This is in combination with a few
other changes (ns_server) to tighten up the requirements and code around
the profiles and ensuring the keys are all being disseminated correctly.

Change-Id: If8af6c8aaf51b71c5817a0cfb0ef85ae242227b8
Reviewed-on: https://review.couchbase.org/c/couchdb/+/225723
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>