MB-66105: [BP] fixup profile loading couchdb
We need access to the profiles from the context of the couchdb node (not
ns_server node) and so we need to load it so it has access to the
persistent_term database as well. This is in combination with a few
other changes (ns_server) to tighten up the requirements and code around
the profiles and ensuring the keys are all being disseminated correctly.

Change-Id: I5ee7517f327e234b613adb6a224bf8c2839a73a1
Reviewed-on: https://review.couchbase.org/c/couchdb/+/226329
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>